refactor: place Next builtin global error adapter
