+++ packages/vitest-plugin-rsc/src/nextjs/src/client/components/builtin/global-error.ts
@@ -1,5 +1,11 @@
 import type { Plugin } from "vite";
 
+// Source: https://github.com/vercel/next.js/blob/ee6e79b1792a4d401ddf2480f40a83549fe8e722/packages/next/src/client/components/builtin/global-error.tsx#L1-L61
+// Adaptation: Next's built-in global-error module is a client component
+// reference in the loader tree. The Vite RSC harness needs a virtual client
+// module with that reference shape; the local test renderer supplies the
+// fallback error UI outside this stub.
+// Begin adapted: Next.js builtin global-error client reference stub
 const virtualNextBuiltinGlobalErrorStubPublicId =
   "virtual:vitest-plugin-rsc/next-builtin-global-error-stub";
 const virtualNextBuiltinGlobalErrorStubId = `\0${virtualNextBuiltinGlobalErrorStubPublicId}`;
@@ -26,3 +32,4 @@ export default function GlobalError() {
     },
   };
 }
+// End adapted
